move namespace memo to own file (#335)
